nifty_agents/requirements_agents.txt
# NIFTY Agent System Requirements
# Install with: pip install -r requirements_agents.txt

# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Data fetching
yfinance>=0.2.31
nsetools>=1.0.11
feedparser>=6.0.10

# Optional: jugaad-data for RBI rates (may have issues on Windows)
# jugaad-data>=0.25

# Google GenAI for LLM
google-generativeai>=0.3.0

# Supabase client
supabase>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For FastAPI testing

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
